# openapi.yaml
openapi: 3.0.3
info:
  title: AI-Powered HRMS API
  description: Next-generation Human Resource Management System with AI-powered resume screening and conversational features
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3001/api
    description: Local development server
  - url: https://your-production-url.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Application:
      type: object
      properties:
        id:
          type: integer
        job_id:
          type: integer
        candidate_name:
          type: string
        candidate_email:
          type: string
          format: email
        candidate_phone:
          type: string
        years_of_experience:
          type: number
        screening_score:
          type: number
          format: float
        status:
          type: string
          enum: [received, screening, screened, shortlisted, interview, rejected, offer, hired]
        applied_at:
          type: string
          format: date-time
paths:
...
  /applications:
    get:
      tags:
        - Applications
      summary: List applications
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: minScore
          in: query
          schema:
            type: number
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                type: object
                properties:
                  applications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Application'
                  pagination:
                    type: object
  /applications/{id}:
    get:
      tags:
        - Applications
      summary: Get application details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                type: object
                properties:
                  application:
                    $ref: '#/components/schemas/Application'
  /applications/screening/{applicationId}/report:
    get:
      tags:
        - Applications
      summary: Get AI screening report
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Screening report
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    type: object
                    properties:
                      applicationId:
                        type: integer
                      candidate:
                        type: object
                      screening:
                        type: object
                        properties:
                          score:
                            type: number
                          status:
                            type: string
                          details:
                            type: object
...